refactor(compiler): use updateConstructor helper in native-constructor.ts
We previously had some exceptionally similar code hanging around. This
just reuses a helper called `updateConstructor` that was introduced a
while back to cut down a bit on code duplication.
